Fix: Only Teacher can see another student

+++ src/Components/info.js
@@ -11,12 +11,14 @@ import history from "../history"
 import {getAllStudent, getAllTeacher} from "../API/service";
 import authenticationService from "../API/authenticationService";
 import {MydModalWithGrid} from "./DialogModal"
+import deleteIcon from '../img/icons8-delete-bin-64.png';
+import detailIcon from '../img/icons8-view-details-64.png';
 
 const Subject = ({ props }) => {
     const [modalShow, setModalShow] = useState(false);
     return (
         <tr className="col-12 row text-center">
-            <td className="col-3 col-md-2">
+            <td className="col-4 col-md-3">
                 <p>{props.name}</p>
             </td>
             <td className="d-none d-sm-block col-md-2">
@@ -25,16 +27,21 @@ const Subject = ({ props }) => {
             <td className="d-none d-sm-block col-md-2">
                 <p>{props.credit}</p>
             </td>
-            <td className="col-3 col-md-2">
+            <td className="col-4 col-md-3">
                 <p>{checkSID(checkCID(props.cid).tid).name}</p>
             </td>
-            <td className="col-6 col-md-4">
-                <button onClick={() => {
-                    courseService.deleteCourse(props.cid)
-                    window.location.reload()
-                }}>X</button>
-                <button onClick={() => setModalShow(true)}>
-                    i
+            <td className="col-4 col-md-2">
+                <button style={{border: "2px solid white",
+                    borderRadius: "25px"}} onClick={async () => {
+                    if (window.confirm('Sure want to delete?')) {
+                        await courseService.deleteCourse(props.cid)
+                        window.location.reload()
+                    }
+                }}><img src={deleteIcon} alt="X" width="20px" height="auto"/></button>
+                <span> </span>
+                <button style={{border: "2px solid white",
+                    borderRadius: "25px"}} onClick={() => setModalShow(true)}>
+                    <img src={detailIcon} alt="i" width="20px" height="auto"/>
                 </button>
                 <MydModalWithGrid show={modalShow} onHide={() => setModalShow(false)} value={props} />
             </td>
@@ -47,6 +54,7 @@ const Subject = ({ props }) => {
         </tr>
     )
 }
+
 
 const handleOnSearch = (string, results) => {
     // getStudentById()
@@ -66,6 +74,7 @@ const handleOnFocus = () => {
 }
 
 export const Info = (props) => {
+    const { isTeacher } = props
     const { id } = props
     const [student, setStudent] = useState()
     const [allStudent, setAllStudent] = useState([])
@@ -114,18 +123,20 @@ export const Info = (props) => {
             <div>
                 <br /> <br />
                 <div className="container header text-center body_font">
-                    <div style={{ width: '380px' }}>
-                        <ReactSearchAutocomplete
-                            items={allStudent}
-                            onSearch={handleOnSearch}
-                            onSelect={handleOnSelect}
-                            onFocus={handleOnFocus}
-                            autoFocus
-                            placeholder="Full Name"
-                            maxResults={2}
-                            inputDebounce={500}
-                        />
-                    </div>
+                    {(isTeacher || student.role === 'Teacher') ?
+                        (<div style={{width: '380px'}}>
+                            <ReactSearchAutocomplete
+                                items={allStudent}
+                                onSearch={handleOnSearch}
+                                onSelect={handleOnSelect}
+                                onFocus={handleOnFocus}
+                                autoFocus
+                                placeholder="Full Name"
+                                maxResults={2}
+                                inputDebounce={500}
+                            />
+                        </div>) : (<></>)
+                    }
                     <h3>{student.role === 'Student'
                         ? `Student` : `Teacher`}'s Infomation</h3>
                     <hr />
@@ -135,13 +146,13 @@ export const Info = (props) => {
                         <tr className="col-12 row">
                             <th className="col-6 col-md-3 col-lg-2">
                                 <img className="student_image"
-                                    src={findCharacterImage(student.name)}
-                                    alt="student" />
+                                     src={findCharacterImage(student.name)}
+                                     alt="student" />
                             </th>
                             <th className="col-6 col-md-6 col-lg-8">
                                 <h5>{student.role === 'Student'
                                     ? `Student` : `Teacher`}'s ID:
-                                        {student.role === 'Student'
+                                    {student.role === 'Student'
                                         ? student.sid : student.tid}</h5>
                                 <h5>Full Name: {student.name}</h5>
                                 <h5>Date of Birth: {student.dob}</h5>
@@ -156,42 +167,37 @@ export const Info = (props) => {
                                 <img className="student_image" src={checkHouseImg(student.house)} alt="house" />
                             </th>
                         </tr>
-                        <tr className="col-12 row text-center">
-                            <td className="col-3 col-md-2">
-                                <h4>Subject</h4>
-                            </td>
-                            <td className="d-none d-sm-block col-md-2">
-                                <h4>Code</h4>
-                            </td>
-                            <td className="d-none d-sm-block col-md-2">
-                                <h4>Credits</h4>
-                            </td>
-                            <td className="col-3 col-md-2">
-                                <h4>Teacher</h4>
-                            </td>
-                            <td className="col-6 col-md-4">
-                                <h4>Action</h4>
-                            </td>
-                            {/* <td className="col-3 col-md-2">
-                                <h4>Room</h4>
-                            </td>
-                            <td className="col-3 col-md-2">
-                                <h4>Time</h4>
-                            </td> */}
-                        </tr>
+                        {student.role === "Student" ?
+                            (<tr className="col-12 row text-center">
+                                <td className="col-4 col-md-3">
+                                    <h4>Subject</h4>
+                                </td>
+                                <td className="d-none d-sm-block col-md-2">
+                                    <h4>Code</h4>
+                                </td>
+                                <td className="d-none d-sm-block col-md-2">
+                                    <h4>Credits</h4>
+                                </td>
+                                <td className="col-4 col-md-3">
+                                    <h4>Teacher</h4>
+                                </td>
+                                <td className="col-4 col-md-2">
+                                    <h4>Action</h4>
+                                </td>
+                            </tr>) : (<></>)
+                        }
                         {allCourse}
                     </table>
                 </div>
                 <br />
-                {id == authenticationService.getId() &&
-                    <div class="row">
-                        <a class="btn btn-info offset-5 col-2" href="/info/score" component={() => <Grades />}>Results</a>
-                    </div>
+                {id === authenticationService.getId() && student.role === "Student" &&
+                <div className="row">
+                    <a className="btn btn-info offset-5 col-2" href={"/info/score"}
+                       component={() => <Grades />}>Results</a>
+                </div>
                 }
-
                 <Footer />
             </div>
         )
     }
-
 }
